refactor(app): remove stripeapp service
