+++ openmeter/appstripe/service.go
@@ -1,17 +1,7 @@
 package appstripe
 
-import (
-	"context"
-
-	appstripeentity "github.com/openmeterio/openmeter/openmeter/appstripe/entity"
-)
-
 type Service interface {
 	AppService
 }
 
-type AppService interface {
-	CreateStripeApp(ctx context.Context, input appstripeentity.CreateAppStripeInput) (appstripeentity.App, error)
-	UpsertStripeCustomerData(ctx context.Context, input appstripeentity.UpsertStripeCustomerDataInput) error
-	DeleteStripeCustomerData(ctx context.Context, input appstripeentity.DeleteStripeCustomerDataInput) error
-}
+type AppService interface{}
